requirements.txt
Django==4.2.7
djangorestframework==3.14.0
drf-yasg==1.21.7
Pillow==10.1.0
coreapi==2.3.3
coreschema==0.0.4
uritemplate==4.1.1
python-decouple==3.8